Fix parsing of file scheme
There was an issue with the file scheme, where it would not recognize
files as paths, since leading slashes would cause the scheme parser to
fail.

The parser is now changed to ignore the number of slashes after the
colon, so paths like `file:///a` are also valid, allowing for proper
file recognition.